refactor + extract pfp component
