+++ packages/client/src/app/components/modals/account/Account.tsx
@@ -218,5 +218,4 @@ export const AccountModal: UIComponent = {
         </ModalWrapper>
       );
     }
-  },
 };
